BACKPORT: FROMLIST: MIPS: Add __cpu_full_name[] to make CPU names more human-readable
In /proc/cpuinfo, we keep "cpu model" as is, since GCC should use it
for -march=native. Besides, we add __cpu_full_name[] to describe the
processor in a more human-readable manner. The full name is displayed
as "model name" in cpuinfo, which is needed by some userspace tools
such as "lscpu" and "gnome-system-monitor".

The CPU frequency in "model name" is the default value (highest), and
there is also a "CPU MHz" whose value can be changed by cpufreq.

This is only used by Loongson now (ICT is dropped in cpu name, and cpu
name can be overwritten by BIOS).

Why drop ICT?
At the beginning, Loongson is designed by ICT, but now all Loongson
processors is designed by "Loongson Technology Corporation Limited"
which is independent from ICT. We have search the code in
https://codesearch.debian.net/, and the result is:

1, GCC searches the "cpu model" in cpuinfo, but it only matches
   "Loongson-2 V0.2" and so on, so drop "ICT" is comfortable;
2, Debian Installer searches the "cpu model" in cpuinfo and matches
   "ICT Loongson", but Yunqiang Su is modifying the code;
3, Valgrind searches the "cpu model" in cpuinfo and matches "ICT
   Loongson", and we are planning to fix it up.

Signed-off-by: Huacai Chen <chenhc@lemote.com>

Link: https://patch.msgid.link/1584006053-28887-1-git-send-email-chenhc@lemote.com
[ Kimi Code: Resolved conflicts on reapplication to
   v7.3-rc1: the 2020 posting predates upstream's
   restructuring of cpu-probe.c, env.c and proc.c;
   re-applied the same semantic changes -
   __cpu_full_name[], cpu_full_name[64],
   mips_cpu_frequency, the model-name/CPU-MHz output in
   /proc/cpuinfo and the BIOS-provided name override -
   against the current code. ]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>