PHYTIUM: net/phytmac: update mdc early to fix PHY ID read issue
The MDC modification in the driver was previously positioned after
mdiobus_register, which caused the MDC not to be properly configured
when reading the PHY ID. This resulted in potential issues during PHY ID
detection.

This commit adjusts the driver by moving the MDC modification logic to
before the call to mdiobus_register, ensuring the MDC is correctly set
up prior to PHY ID reading operations.

Signed-off-by: liutianyu1250 <liutianyu1250@phytium.com.cn>

Link: https://gitee.com/phytium_embedded/phytium-linux-kernel/commit/7c9ca9af7d15d02f20347c3030f179881f3eb596
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>