BACKPORT: FROMLIST: rust: export BINDGEN_TARGET from a separate Makefile
A subsequent commit will add a new function `bindgen-option` to
`scripts/Kconfig.include`. The bindgen backend requires the `--target`
option for cross compiling, but variable `BINDGEN_TARGET` in
`rust/Makefile` cannot be exported to `scripts/Kconfig.include`.

Therefore, move this variable to a separate new `Makefile.rust` file and
include it from `scripts/Makefile` to make the exported variable
available for use in Kconfig. Place the include in the `need-compiler`
branch to avoid including it in irrelevant make targets.

Since the new file name is `Makefile.rust`, it matches an existing
MAINTAINERS rule `scripts/*rust*`, so no modification to the MAINTAINERS
file is needed.

Signed-off-by: Asuna Yang <xinrui.riscv@isrc.iscas.ac.cn>
Acked-by: Nathan Chancellor <nathan@kernel.org>
Acked-by: Conor Dooley <conor.dooley@microchip.com>

Link: https://patch.msgid.link/20251230-gcc-rust-v5-v6-1-2ac86ba728c8@isrc.iscas.ac.cn
[ Kimi Code: on reapplication to v7.3-rc1, the BINDGEN_TARGET
   block in rust/Makefile also carried upstream's newer s390, i386
   and powerpc targets; moved the full block, including those
   targets, to scripts/Makefile.rust. ]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>